Moved firebase to util
