+++ src/components/Place/PlaceTable.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { usePlaces } from "../Places/Places";
-import { useFirebase } from "../Firebase";
+import { useFirebase } from '../../util/Firebase';
 import { useAuth } from "../Session/UserAuth";
 
 import "./PlaceTable.css";
